Add join and waitlist frontend
`join.html` contains the code for displaying the frontend for the join page that gets served at `/join-waitlist`. The `/add` endpoint has been changed to serve the `waitlist.html` file after adding the person to the database. A repeated attempt to add an existing member to the waitlist (aka reloading the page) will not cause issues.